+++ var/starter_kits/italiagov/src/js/custom/custom.js
@@ -7,3 +7,4 @@
 //import './example'
 //import './example-alert-event'
 //import './example-bootstrap-italia-modules'
+//import './icons'